requirements.txt
pandas
yfinance
investpy
numpy
matplotlib
seaborn
requests
datetime
google-cloud-bigquery
google-cloud-storage